Update tests for CustomElement.
