--- tests/unit/Widget-attr.js
@@ -1,125 +1,0 @@
-define([
-	"intern!object",
-	"intern/chai!assert",
-	"delite/register",
-	"dojo/sniff",
-	"delite/Widget",
-	"dojo/domReady!"
-], function (registerSuite, assert, register, has, Widget) {
-	var container;
-	registerSuite({
-		name: "Widget-attr",
-		setup: function () {
-			container = document.createElement("div");
-			document.body.appendChild(container);
-		},
-		"declarative" : function () {
-			global = 0;		// create intentional global
-			global2 = { text: "global var" };
-			var Declarative = register("test-declarative", [HTMLElement, Widget], {
-				boolProp: false,
-				numProp: 0,
-				stringProp: "",
-				funcProp : function () { },
-				objProp1: { },
-				objProp2: { }
-			});
-			container.innerHTML +=
-				"<test-declarative id='d' boolProp='boolProp' numProp='5' stringProp='hello' funcProp='global=123;' objProp1='foo:1,bar:2' objProp2='global2'/>";
-			var d = document.getElementById("d");
-			register.upgrade(d);
-			assert.isTrue(d.boolProp, "d.boolProp");
-
-			assert.strictEqual(5, d.numProp, "d.numProp");
-			assert.strictEqual("hello", d.stringProp, "d.stringProp");
-			d.funcProp();
-			assert.strictEqual(123, global, "d.funcProp() executed");
-			assert.strictEqual(1, d.objProp1.foo, "d.objProp1.foo");
-			assert.strictEqual(2, d.objProp1.bar, "d.objProp1.bar");
-			assert.strictEqual("global var", d.objProp2.text, "d.objProp2.text");
-		},
-
-		// tabIndex is problematic, see https://github.com/ibm-js/delite/issues/34.
-		"specialNames" : function () {
-			// Test when tabIndex is declared top level, in the props passed to register().
-			// TODO: enable when https://github.com/uhop/dcl/issues/9 is fixed
-			/*
-			var SpecialNames = register("test-special-names", [HTMLElement, Widget], {
-				tabIndex: "0",
-
-				postCreate: function () {
-					this.watch("tabIndex", function(name, o, n){
-						this.watchedTabIndex = n;
-					});
-				}
-			});
-			var widget = new SpecialNames({ });
-			widget.tabIndex = "3";
-			document.body.appendChild(widget);
-			widget.startup();
-			assert.equal("3", widget.watchedTabIndex, "watch fired on widget");
-			*/
-
-			// And test when tabIndex is declared in a mixin.
-			var SpecialNamesMixin = register.dcl(Widget, {
-				tabIndex: "0",
-
-				postCreate: function () {
-					this.watch("tabIndex", function(name, o, n){
-						this.watchedTabIndex = n;
-					});
-				}
-			});
-			var SpecialExtendedWidget = register("test-special-names-extended", [HTMLElement, SpecialNamesMixin], {
-				tabIndex: "0",
-				value: "0",
-				isrange: false,
-				isbool: false,
-
-				postCreate: function () {
-					this.watch("tabIndex", function(name, o, n){
-						this.watchedTabIndex = n;
-					});
-				}
-			});
-
-			var extended = new SpecialExtendedWidget({ });
-			extended.tabIndex = "5";
-			document.body.appendChild(extended);
-			extended.startup();
-			assert.equal("5", extended.watchedTabIndex, "watch fired on extended");
-
-			// And also test for declarative widgets, to make sure the tabIndex property is
-			// removed from the root node, to prevent an extra tab stop
-			container.innerHTML += "<test-special-names-extended id=specialNames value=5 isrange isbool tabIndex=8/>";
-			var declarative = document.getElementById("specialNames");
-			register.upgrade(declarative);
-			assert.isFalse(declarative.hasAttribute("tabindex"), "tabindex attr removed");
-			assert.isTrue(declarative.isrange, "isrange set");
-			assert.isTrue(declarative.isbool, "isbool set");
-			assert.strictEqual("5", declarative.value, "value");
-
-			// Finally, test when the widget prototype doesn't declare tabIndex at all.
-			// Then the widget should just act like a simple <div>, passing tabIndex through to root node.
-			var SimpleWidget = register("simple-widget", [HTMLElement, Widget], { });
-			var simple = new SimpleWidget({ tabIndex: 5 });
-			document.body.appendChild(simple);
-			simple.startup();
-
-			// make sure that tabIndex was correctly set
-			assert.strictEqual("5", simple.getAttribute("tabindex"), "programmatic set");
-
-			// And also test for declarative widgets, to make sure the tabIndex property is
-			// removed from the root node, to prevent an extra tab stop
-			container.innerHTML += "<simple-widget id=simple tabIndex=8>";
-			var simpleDeclarative = document.getElementById("simple");
-			register.upgrade(simpleDeclarative);
-
-			// make sure that tabIndex wasn't unset
-			assert.strictEqual("8", simpleDeclarative.getAttribute("tabindex"), "declarative set");
-		},
-		teardown : function () {
-			container.parentNode.removeChild(container);
-		}
-	});
-});
