+++ tests/unit/all.js
@@ -3,16 +3,13 @@ define([
 	"./Destroyable",
 	"./DisplayContainer",
 	"./register",
-	"./Widget-attr",
+	"./CustomElement",
+	"./Widget",
 	"./handlebars",
-	"./Widget-lifecycle",
-	"./Widget-placeAt",
-	"./Widget-utility",
 	"./HasDropDown",
 	"./CssState",
 	"./Container",
 	"./a11y",
-	"./Widget-on",
 	"./Stateful",
 	"./Selection",
 	"./StoreMap",
